+++ collective/tablepage/browser/table.py
@@ -105,7 +105,8 @@ class TableViewView(BrowserView):
                 continue
             results.append(dict(label=tablepageMessageFactory(d['label'].decode('utf-8')),
                                 description=tablepageMessageFactory(d.get('description', '').decode('utf-8')),
-                                classes='coltype-%s' % self.ploneview.normalizeString(d['type']),
+                                classes='coltype-%s col-%s' % (self.ploneview.normalizeString(d['type']),
+                                                               self.ploneview.normalizeString(d['id']),),
                                 ))
         return results
 
@@ -225,7 +226,9 @@ class TableViewView(BrowserView):
                         write_attempt = True
                         logger.debug("Cache miss (%s)" % conf['id'])
                 row['cols'].append({'content': output,
-                                    'classes': 'coltype-%s' % self.ploneview.normalizeString(col_type)})
+                                    'classes': 'coltype-%s col-%s' % (self.ploneview.normalizeString(col_type),
+                                                                      self.ploneview.normalizeString(conf['id'])
+                                                                      )})
             rows.append(row)
             index += 1
             if write_attempt: